Make linter schema a separate json file
