+++ cli/linter/linter.go
@@ -1,4 +1,4 @@
-package lint
+package linter
 
 import (
 	"encoding/json"
